+++ app/apis/getPostsAmount.ts
